MultiPuz-Game Problems
I can't see it on the build screen when I drag an drop the parts, but when I place it in the right place on the Home screen, I get the image on the build screen, but when I do the same on the build screen, there are no changes on the Home screen. I don't know how to solve this problem, I would appreciate it if you could help.